불필요한 컴포넌트 정리
